Fix: get the time when start, pause, resume and stop

+++ app/models/record_stopwatch.py
@@ -10,5 +10,5 @@ class RecordStopwatch(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     task_id = Column(UUID(as_uuid=True), ForeignKey('task.id'), nullable=False)
     quantity = Column(Float, nullable=False)
-    accumulated_duration = Column(BigInteger, default=0, nullable=False) # in mili-seconds
+    accumulated_duration = Column(Float, default=0, nullable=False) # in hours
     task = relationship('Task', back_populates='record_stopwatch')
